Initial work to convert the layout to feel more "inboxy"
